get valid downloaded number from new left variable

+++ macosx/Torrent.h
@@ -176,7 +176,7 @@
 
 - (float)       downloadRate;
 - (float)       uploadRate;
-- (float)       downloadedValid;
+- (uint64_t)       downloadedValid;
 - (uint64_t)    downloadedTotal;
 - (uint64_t)    uploadedTotal;
 - (float)       swarmSpeed;
